+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Provide return types
 - New test for changing the default user agent
 - Support for HEAD requests via head() method
+- New test for head() method
 
 ## 0.0.4 - 2023-04-12
 - Github Actions for tests