+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support for HEAD requests via head() method
 - New test for head() method
 - Support for HEAD requests via h() method
+- New test for h() method
 
 ## 0.0.4 - 2023-04-12
 - Github Actions for tests